requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
openai==1.55.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
allure-pytest==2.13.5
ruff==0.11.12
